Made Hash Tables's fields private
